Android: WEbView hook to set tab index

+++ platform/android/Rhodes/src/com/rhomobile/rhodes/extmanager/IRhoWebView.java
@@ -22,7 +22,7 @@ public interface IRhoWebView {
     void loadData(String data, String mime, String encoding);
     void loadDataWithBaseURL(String baseUrl, String data, String mimeType, String encoding, String historyUrl);
     void stopLoad();
-    void setZoom(int scale);
+    void setZoom(double scale);
     void setTextZoom(int scale);
     String getEngineId();
     void onPause();
